Neatened up uleb128/sleb128 values, fixed a bug that would cause a crash if a file was ended by a Uleb value, added some test coverage

+++ src/parser/encoded_value.rs
@@ -86,7 +86,7 @@ named!(pub parse_encoded_array_item<&[u8], EncodedArrayItem>,
 
 #[derive(Debug, PartialEq)]
 pub struct EncodedArrayItem {
-    size: u64,
+    size: u32,
     values: Vec<EncodedValue>
 }
